Update Main.ino
Now it's working. To use:
Plug into power.
Type time in military format (i.e if PM, add 12.)
If it's before 10 AM, you must write a zero before the hour (9:50 AM means you type 0959, not 959). Times are preprogrammed so morning starts at 6 am and goes to noon, evening is from 8 to midnight. Morning is tray 1, evening is tray 2. 
The color code is that green means morning is ready, yellow means evening is ready, and red means it cannot dispense pills.